@@ -5,4 +5,4 @@ build-backend = "setuptools.build_meta"
 [tool.setuptools_scm]
 
 [tool.black]
-target-version = ['py36']
+target-version = ['py37']
