@@ -19,17 +19,17 @@ classifiers =
     Operating System :: MacOS
     Operating System :: Microsoft :: Windows
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
 
 [options]
 package_dir =
     =src
 packages = find:
-python_requires = >=3.6, <3.11
+python_requires = >=3.7, <3.12
 
 [options.packages.find]
 where = src
@@ -37,6 +37,7 @@ where = src
 [options.extras_require]
 # Dependencies for testing.
 test =
+    covdefaults
     pytest
     pytest-cov
 
@@ -46,26 +47,19 @@ test =
 testpaths = tests
 
 [coverage:run]
-branch = True
+plugins = covdefaults
 source = docstring_inheritance
-
-[coverage:report]
-exclude_lines =
-    pragma: no cover
 
 [flake8]
 # See http://www.pydocstyle.org/en/latest/error_codes.html for more details.
 # https://github.com/PyCQA/flake8-bugbear#how-to-enable-opinionated-warnings
 ignore =
-    E501
-    # no docstring for standard and private methods
+    # No docstring for standard and private methods.
     D105
-    E203
     W503
 # select = B,C,D,E,F,G,N,T,W
-select = B,C,E,F,G,N,T,W
-# settings for compatibility with black, see
-# https://github.com/psf/black/blob/master/docs/compatible_configs.md#flake8
-# https://black.readthedocs.io/en/stable/the_black_code_style.html?highlight=bugbear#line-length
+select = B,B950,C,E,F,G,N,T,W
+# Settings for compatibility with black, see black docs
 max-line-length = 88
 docstring-convention = google
+extend-ignore = E203, E501
