@@ -17,9 +17,12 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
+from __future__ import annotations
+
 import pytest
-from docstring_inheritance import GoogleDocstringProcessor
 from test_base_processor import _test_parse_sections
+
+from docstring_inheritance import GoogleDocstringProcessor
 
 
 @pytest.mark.parametrize(
